--- EnkuToolkit.Wpf.Demo/Services/INavigationService.cs
@@ -1,12 +1,0 @@
-﻿namespace EnkuToolkit.Wpf.Demo.Services;
-
-using System;
-
-public interface INavigationService
-{
-    bool Navigate(string uriStr, object? extraData = null);
-
-    void GoForward();
-
-    void GoBack();
-}
